cleanup(cmd,pkg): moved `enum` to its own package.
Moreover, moved logLevel and logFormat enums under pkg/output.
